requirements.txt
llama-cpp-python>=0.2.71
huggingface-hub>=0.22.2
python-dotenv>=1.0.0
openai